Add drink method and AlreadyUsedException
Now Potions have a "drink" method that will, ehm, drink the potions.
Heal Potions have their drink method which will affect an entity
A new class (AlreadyUsedException), subclass of error, was created to
be thrown in case an Item (such as a potion) that is consumable already
has been consumed.